Refactoring some tests, created the ItemApplicationService and respective tests

+++ simpleCrud/simple-crud-core/src/main/java/com/portofolio/demo/infrastructure/persistence/item/ItemRepositoryServiceImpl.java
@@ -1,6 +1,7 @@
 package com.portofolio.demo.infrastructure.persistence.item;
 
 import com.portofolio.demo.domain.item.Item;
+import com.portofolio.demo.shared.errors.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -31,6 +32,10 @@ public class ItemRepositoryServiceImpl implements ItemRepositoryService {
     public void deleteById(Long id) {
         if (id == null) {
             throw new IllegalArgumentException("Id can not be null");
+        }
+
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Item not found with id: " + id, null);
         }
 
         repository.deleteById(id);
